requirements.micropython.txt
micropython-ulogging
micropython-itertools
micropython-urequests
micropython-uasyncio
